requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==41.0.7
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.0.1
WTForms==3.1.1
